tests/WeKan.Application.UnitTests/Commands/DeleteBoard/DeleteBoardAuthorizerTests.cs
using Moq;
using System;
using System.Threading;
using System.Threading.Tasks;
using WeKan.Application.Commands.DeleteBoard;
using WeKan.Application.Common.Interfaces;
using WeKan.Domain.Boards;
using Xunit;

namespace WeKan.Application.UnitTests.Commands.DeleteBoard
{
    public class DeleteBoardAuthorizerTests
    {
        private readonly Mock<ICurrentUserPermissionService> _currentUserPermissionService;

        public DeleteBoardAuthorizerTests()
        {
            _currentUserPermissionService = new Mock<ICurrentUserPermissionService>();
        }

        [Fact]
        public void Ctor_ICurrentUserPermissionServiceNull_ThrowsArgumentNullException()
        {
            Assert.Throws<ArgumentNullException>(() => new DeleteBoardAuthorizer(null));
        }

        [Fact]
        public async Task Authorise_CallsHasPermissionForBoard_CanDeleteBoard()
        {
            var permission = BoardUserPermission.CAN_DELETE_BOARD;
            var expectedResult = true;
            var cancellationToken = new CancellationToken();
            _currentUserPermissionService
                .Setup(s => s.HasPermissionForBoard(
                    It.IsAny<int>(),
                    It.Is<BoardUserPermission>(p => p == permission),
                    cancellationToken))
                .ReturnsAsync(expectedResult);

            var authorizer = new DeleteBoardAuthorizer(_currentUserPermissionService.Object);

            var result = await authorizer.Authorise(new DeleteBoardCommand(1), cancellationToken);

            Assert.Equal(expectedResult, result);
            _currentUserPermissionService
                .Verify(s => s.HasPermissionForBoard(It.IsAny<int>(), It.Is<BoardUserPermission>(p => p == permission), cancellationToken), Times.Once);
        }
    }
}
